alx_travel_app/requirements.txt
Django
djangorestframework
mysqlclient
django-cors-headers
drf-yasg
django-environ
celery
rabbitmq
